requirements.txt
altair==5.5.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.10.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
attrs==25.3.0
babel==2.17.0
beautifulsoup4==4.13.4
bleach==6.2.0
blinker==1.9.0
bs4==0.0.2
cachetools==6.2.0
certifi==2025.10.5
cffi==1.17.1
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
debugpy==1.8.16
decorator==5.2.1
defusedxml==0.7.1
distro==1.9.0
dnspython==2.8.0
docx==0.2.4
email-validator==2.3.0
executing==2.2.0
fastapi==0.124.4
fastapi-cli==0.0.16
fastapi-cloud-cli==0.7.0
fastar==0.8.0
fastjsonschema==2.21.1
filetype==1.2.0
fonttools==4.59.1
fqdn==1.5.1
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.45.0
google-genai==1.56.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.10
ipykernel==6.30.1
ipython==9.4.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.2
json5==0.12.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.6
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.16.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
kiwisolver==1.4.9
langchain==1.2.0
langchain-core==1.2.2
langchain-google-genai==4.1.1
langchain-ollama==1.0.1
langgraph==1.0.5
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.5
langgraph-sdk==0.3.0
langsmith==0.5.0
lark==1.2.2
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.5
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.3
narwhals==2.5.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
notebook==7.4.5
notebook_shim==0.2.4
numpy==2.3.2
ollama==0.6.1
orjson==3.11.5
ormsgpack==1.12.1
outcome==1.3.0.post0
overrides==7.7.0
packaging==25.0
pandas==2.3.1
pandocfilters==1.5.1
parso==0.8.4
pillow==11.3.0
platformdirs==4.3.8
prometheus_client==0.22.1
prompt_toolkit==3.0.51
protobuf==6.32.1
psutil==7.0.0
pure_eval==0.2.3
py4j==0.10.9.9
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
pyparsing==3.2.3
PyPDF2==3.0.1
PySocks==1.7.1
pyspark==4.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-json-logger==3.3.0
python-multipart==0.0.21
pytz==2025.2
pywin32==311
pywinpty==3.0.0
PyYAML==6.0.2
pyzmq==27.0.1
referencing==0.36.2
requests==2.32.4
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.2.0
rich-toolkit==0.17.0
rignore==0.7.6
rpds-py==0.27.0
rsa==4.9.1
scikit-learn==1.7.2
scipy==1.16.2
seaborn==0.13.2
selenium==4.38.0
Send2Trash==1.8.3
sentry-sdk==2.48.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
spark-nlp==6.2.3
stack-data==0.6.3
starlette==0.50.0
streamlit==1.50.0
tenacity==9.1.2
terminado==0.18.1
threadpoolctl==3.6.0
tinycss2==1.4.0
toml==0.10.2
tornado==6.5.2
traitlets==5.14.3
trio==0.32.0
trio-websocket==0.12.2
typer==0.20.0
types-python-dateutil==2.9.0.20250809
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uri-template==1.3.0
urllib3==2.5.0
uuid_utils==0.12.0
uvicorn==0.38.0
watchdog==6.0.0
watchfiles==1.1.1
wcwidth==0.2.13
webcolors==24.11.1
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
widgetsnbextension==4.0.14
wsproto==1.2.0
xxhash==3.6.0
zstandard==0.25.0
